+++ tests/extra_statics_tools/phan/constants.php
@@ -1,0 +1,5 @@
+<?php
+
+if ( ! defined( 'HOUR_IN_SECONDS.' ) ) {
+	define('HOUR_IN_SECONDS', 9999999 );
+}